requirements.txt

torch
numpy
matplotlib
pyyaml
pytest-shutil
scipy
ResNet